+++ back-end/src/main/java/com/apple/shop/orders/OrdersDTO.java
@@ -1,7 +1,6 @@
 package com.apple.shop.orders;
 
-import com.apple.shop.cart.Cart; // CartItem 클래스 import
-import lombok.Data;
+import com.apple.shop.cart.Cart;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -15,6 +14,8 @@ public class OrdersDTO {
     private String orderMemo;
     private String deliveryAddr;
     private String recipientPhone;
-
-
+    private String recipientName;
+    private String payMethod;
+    private Integer payAmount;
+    private Long storeIdx;
 }