+++ back-end/src/main/java/com/apple/shop/storeMember/StoreInfoDTO.java
@@ -1,0 +1,18 @@
+package com.apple.shop.storeMember;
+
+import lombok.Getter;
+import lombok.Setter;
+
+/**
+ * 가게 정보 저장 요청 DTO
+ */
+@Getter
+@Setter
+public class StoreInfoDTO {
+    private String storeName;
+    private String storePhone;
+    private String storeAddr;
+    private String storeNotice;
+    private String businessHours;
+    private String storeImg;
+}